Fix breakage of --overwrite without an existing installation

+++ setup.sh
@@ -234,7 +234,7 @@ create_env() {
 
     # Move any existing environment file out of the way
     TIMESTAMP_NOW=$(date +'%Y.%m.%d-%H.%M')
-    mv -f "${REDASH_BASE_PATH}/env" "${REDASH_BASE_PATH}/env.old.${TIMESTAMP_NOW}"
+    mv -f "${REDASH_BASE_PATH}/env" "${REDASH_BASE_PATH}/env.old-${TIMESTAMP_NOW}"
   fi
 
   echo "Generating brand new environment file"
@@ -257,7 +257,7 @@ setup_compose() {
 
   cd "$REDASH_BASE_PATH"
   GIT_BRANCH="${REDASH_BRANCH:-master}" # Default branch/version to master if not specified in REDASH_BRANCH env var
-  if [ "x$OVERWRITE" = "xyes" ]; then
+  if [ "x$OVERWRITE" = "xyes" -a -e compose.yaml ]; then
     mv -f compose.yaml compose.yaml.old
   fi
   curl -fsSOL https://raw.githubusercontent.com/getredash/setup/"$GIT_BRANCH"/data/compose.yaml
